Fixed Sanity Test Issues
